wiki client modularized from github.com/dobbs/wiki-spike-css/pwa
Pretty happy with the improvement to library.js compared to pwa. In
particular, we combined the link parsing with DOMPurify into a single
function replacement for Observable's html tagged template literal.

Also happy with the inspiration to use a custom event for wiki
links. I have not yet created a link handler, but I already like using
an event to keep the parsing and security separate from the link
handling.