use a local copy of Observable's htl library
Also insert our DOMPurify sanitizer into all uses of htl or html.

This implementation has some complexity. The published htl library we
have copied locally is an AMD style module, not an ESM module. So we
use the window.htl object that gets created as a side-effect of
the dynamic import().

The other thing we've done intentionally is to run the import()s in
parallel (Using Promise.all()) when loading DOMPurify and the htl
library;